+++ src/main/java/ch/ethz/matsim/av/generator/PopulationDensityGenerator.java
@@ -25,7 +25,7 @@ import java.util.*;
 public class PopulationDensityGenerator implements AVGenerator {
     private final Random random;
     private final long numberOfVehicles;
-    private final long numberOfSeats;
+    private final int numberOfSeats;
 	private long generatedNumberOfVehicles = 0;
 
     private final String prefix;
@@ -34,7 +34,7 @@ public class PopulationDensityGenerator implements AVGenerator {
     private Map<Link, Double> cumulativeDensity = new HashMap<>();
 
     public PopulationDensityGenerator(AVGeneratorConfig config, Network network, Population population,
-									  ActivityFacilities facilities, int randomSeed, long numberOfSeats) {
+									  ActivityFacilities facilities, int randomSeed, int numberOfSeats) {
         this.random = new Random(randomSeed);
         this.numberOfVehicles = config.getNumberOfVehicles();
         this.numberOfSeats = numberOfSeats;
@@ -108,7 +108,7 @@ public class PopulationDensityGenerator implements AVGenerator {
         @Override
         public AVGenerator createGenerator(AVGeneratorConfig generatorConfig) {
             int randomSeed = Integer.parseInt(generatorConfig.getParams().getOrDefault("randomSeed", "1234"));
-            long numberOfSeats = Long.parseLong(generatorConfig.getParams().getOrDefault("numberOfSeats", "4"));
+            int numberOfSeats = Integer.parseInt(generatorConfig.getParams().getOrDefault("numberOfSeats", "4"));
             return new PopulationDensityGenerator(generatorConfig, network, population, facilities, randomSeed, numberOfSeats);
         }
     }
